Add chainName to walletStatus type
