+++ src/hooks/useContractNetworksOnly.ts
@@ -2,9 +2,9 @@ import { useEffect } from 'react';
 
 import { useWeb3React } from '@web3-react/core';
 import { vaultsSupportedChainIds } from 'common-lib/chains';
-import { switchNetwork } from 'lib/web3-helpers';
 
 import { useApeSnackbar } from 'hooks';
+import { switchNetwork } from 'utils/provider';
 
 export default function useContractNetworksOnly() {
   const { chainId } = useWeb3React();
